feat: move `CopyButton` to module
